Update projects to use sdk-style
- Gtk2 and XamMac still need the old style, ugh.
