+++ Eto.Gl.Wpf/Eto.Gl.Wpf.csproj
@@ -1,0 +1,24 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <TargetFramework>net45</TargetFramework>
+  </PropertyGroup>
+  <ItemGroup>
+    <Compile Include="..\Eto.Gl.WinForms\Win32Helper.cs" Link="Win32Helper.cs" />
+    <Compile Include="..\Eto.Gl.WinForms\WinGLUserControl.cs" Link="WinGLUserControl.cs" />
+    <Compile Include="..\Eto.Gl.WinForms\WinGLUserControl.designer.cs" Link="WinGLUserControl.designer.cs" />
+  </ItemGroup>
+  <ItemGroup>
+    <PackageReference Include="Eto.Platform.Wpf" Version="2.4.1" />
+    <PackageReference Include="OpenTK" Version="3.0.1" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\Eto.Gl\Eto.Gl.csproj" />
+  </ItemGroup>
+  <ItemGroup>
+    <Reference Include="PresentationCore" />
+    <Reference Include="PresentationFramework" />
+    <Reference Include="System.Windows.Forms" />
+    <Reference Include="WindowsBase" />
+    <Reference Include="WindowsFormsIntegration" />
+  </ItemGroup>
+</Project>
