+++ Eto.Gl.sln
@@ -3,29 +3,29 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.28917.181
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Eto.Gl.Gtk2", "Eto.Gl.Gtk2\Eto.Gl.Gtk2.csproj", "{06F7AC59-0E70-4646-BBFA-C105D9664FD0}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Eto.Gl.Gtk2", "Eto.Gl.Gtk\Eto.Gl.Gtk2.csproj", "{06F7AC59-0E70-4646-BBFA-C105D9664FD0}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Eto.Gl.WinForms", "Eto.Gl.WinForms\Eto.Gl.WinForms.csproj", "{A0DCCC13-936C-4D5A-AD8C-ADFF75816A8F}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Eto.Gl.Mac", "Eto.Gl.Mac\Eto.Gl.Mac.csproj", "{C6D7920E-3FB5-4DAD-80F1-2B76158D493F}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Eto.Gl.Mac64", "Eto.Gl.Mac\Eto.Gl.Mac64.csproj", "{C6D7920E-3FB5-4DAD-80F1-2B76158D493F}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestEtoGl.Mac", "TestEtoGl.Mac\TestEtoGl.Mac.csproj", "{5DD54367-9089-4371-BF0A-1BE772B623E6}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestEtoGl.Mac64", "TestEtoGl.Mac\TestEtoGl.Mac64.csproj", "{5DD54367-9089-4371-BF0A-1BE772B623E6}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestEtoGl", "TestEtoGl\TestEtoGl.csproj", "{66C691D6-557F-4477-BF96-FA5BC92C715B}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestEtoGl.WinForms", "TestEtoGl.WinForms\TestEtoGl.WinForms.csproj", "{2A29DA7C-7CB0-458A-9EA8-85811AD952E1}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestEtoGl.XamMac", "TestEtoGl.XamMac\TestEtoGl.XamMac.csproj", "{E201F515-1163-401A-ABC0-17A0E6A85764}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestEtoGl.XamMac2", "TestEtoGl.XamMac\TestEtoGl.XamMac2.csproj", "{E201F515-1163-401A-ABC0-17A0E6A85764}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Eto.Gl.XamMac", "Eto.Gl.XamMac\Eto.Gl.XamMac.csproj", "{D260DEBB-3D35-46A4-AB23-F2A19CC10E28}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Eto.Gl.XamMac2", "Eto.Gl.Mac\Eto.Gl.XamMac2.csproj", "{D260DEBB-3D35-46A4-AB23-F2A19CC10E28}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Eto.Gl", "Eto.Gl\Eto.Gl.csproj", "{510CE98D-2169-48A0-889C-E743417EE5E1}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestEtoGl.Gtk2", "TestEtoGl.Gtk2\TestEtoGl.Gtk2.csproj", "{02F28D06-103A-4800-9E40-E9717D77449F}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestEtoGl.WPF_WinformsHost", "TestEtoGl.WPF_WinformsHost\TestEtoGl.WPF_WinformsHost.csproj", "{B1EB0BB0-C662-4314-A3D3-20E1AD0F98D6}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestEtoGl.Wpf", "TestEtoGl.Wpf\TestEtoGl.Wpf.csproj", "{B1EB0BB0-C662-4314-A3D3-20E1AD0F98D6}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Eto.Gl.WPF_WinformsHost", "Eto.Gl.WPF_WinformsHost\Eto.Gl.WPF_WinformsHost.csproj", "{43C9116A-2F41-4F57-88E5-6B49BFB7881B}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Eto.Gl.Wpf", "Eto.Gl.Wpf\Eto.Gl.Wpf.csproj", "{43C9116A-2F41-4F57-88E5-6B49BFB7881B}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Build", "Build", "{A4E13914-032E-461F-BC5C-5A88D42AE19D}"
 	ProjectSection(SolutionItems) = preProject
@@ -92,11 +92,5 @@ Global
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {7B6C542A-8C6C-4200-8971-67942032E485}
-	EndGlobalSection
-	GlobalSection(MonoDevelopProperties) = preSolution
-		Policies = $0
-		$0.TextStylePolicy = $1
-		version = 0.24
-		StartupItem = SharpFlame.Gui.Linux\SharpFlame.Gui.Linux.csproj
 	EndGlobalSection
 EndGlobal
