+++ TestEtoGl.Mac/TestEtoGl.Mac64.csproj
@@ -2,13 +2,14 @@
   <PropertyGroup>
     <TargetFramework>net462</TargetFramework>
     <OutputType>WinExe</OutputType>
+    <DefineConstants>MONOMAC</DefineConstants>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="Eto.Platform.Windows" Version="2.5.0-beta.1" />
+    <PackageReference Include="Eto.Platform.Mac64" Version="2.5.0-beta.1" />
     <PackageReference Include="OpenTK" Version="3.0.1" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\TestEtoGl\TestEtoGl.csproj" />
-    <ProjectReference Include="..\Eto.Gl.WinForms\Eto.Gl.WinForms.csproj" />
+    <ProjectReference Include="..\Eto.Gl.Mac\Eto.Gl.Mac64.csproj" />
   </ItemGroup>
 </Project>
