+++ TestEtoGl.Wpf/TestEtoGl.Wpf.csproj
@@ -4,11 +4,11 @@
     <OutputType>WinExe</OutputType>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="Eto.Platform.Windows" Version="2.5.0-beta.1" />
+    <PackageReference Include="Eto.Platform.Wpf" Version="2.5.0-beta.1" />
     <PackageReference Include="OpenTK" Version="3.0.1" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\TestEtoGl\TestEtoGl.csproj" />
-    <ProjectReference Include="..\Eto.Gl.WinForms\Eto.Gl.WinForms.csproj" />
+    <ProjectReference Include="..\Eto.Gl.Wpf\Eto.Gl.Wpf.csproj" />
   </ItemGroup>
 </Project>
